Fix missing exit code on build script failure (#523)
Currently the build command always exits with a `0` code when
a compilation error is encountered which prevents compilation 
failures failing builds in CI environments.

>The `err` object will not include compilation errors and those 
>must be handled separately using `stats.hasErrors()` [...]. The 
>`err` object will only contain webpack-related issues, such as
>misconfiguration, etc.
>[webpack node API documentation][3]

This PR sets the [exit code][1] to 1 when an error is encountered 
but lets the process exit naturally rather than a force exit, which 
[can be problematic][2] when async process are in play.

This has benefits over for exiting the process i.e. not exiting when
watching. It also doesn't require adding additional or cli flags.

Additionally this preserves non-zero exit codes if they've already
been set to preserve BC.

[1]: https://nodejs.org/api/process.html#process_process_exitcode
[2]: https://nodejs.org/api/process.html#process_process_exit
[3]: https://webpack.js.org/api/node/#webpack-